+++ test/integration/ddl_sharding_key_test.lua
@@ -706,6 +706,8 @@ pgroup.test_update_cache_with_incorrect_key = function(g)
     -- records for all spaces exist
     sharding_key_as_index_obj = helpers.get_sharding_key_cache(g.cluster)
     t.assert_equals(sharding_key_as_index_obj, {
+        customers_G_func = {parts = {{fieldno = 1}}},
+        customers_body_func = {parts = {{fieldno = 1}}},
         customers_age_key = {parts = {{fieldno = 4}}},
         customers_name_age_key_different_indexes = {parts = {{fieldno = 3}, {fieldno = 4}}},
         customers_name_age_key_three_fields_index = {parts = {{fieldno = 3}, {fieldno = 4}}},
@@ -730,6 +732,8 @@ pgroup.test_update_cache_with_incorrect_key = function(g)
     -- other records for correct spaces exist in cache
     sharding_key_as_index_obj = helpers.get_sharding_key_cache(g.cluster)
     t.assert_equals(sharding_key_as_index_obj, {
+        customers_G_func = {parts = {{fieldno = 1}}},
+        customers_body_func = {parts = {{fieldno = 1}}},
         customers_age_key = {parts = {{fieldno = 4}}},
         customers_name_age_key_different_indexes = {parts = {{fieldno = 3}, {fieldno = 4}}},
         customers_name_age_key_three_fields_index = {parts = {{fieldno = 3}, {fieldno = 4}}},
@@ -753,6 +757,8 @@ pgroup.test_update_cache_with_incorrect_key = function(g)
     -- other records for correct spaces exist in cache
     sharding_key_as_index_obj = helpers.get_sharding_key_cache(g.cluster)
     t.assert_equals(sharding_key_as_index_obj, {
+        customers_G_func = {parts = {{fieldno = 1}}},
+        customers_body_func = {parts = {{fieldno = 1}}},
         customers_age_key = {parts = {{fieldno = 4}}},
         customers_name_age_key_different_indexes = {parts = {{fieldno = 3}, {fieldno = 4}}},
         customers_name_age_key_three_fields_index = {parts = {{fieldno = 3}, {fieldno = 4}}},